refactor(web): rename routes-rex/shared.ts to rex-route-helpers.ts
Eliminates naming collision between routes-rex/shared.ts (route-level
utilities) and server/shared-types.ts (cross-zone type bridge). Both
used "shared" for structurally different purposes.

All 8 import sites updated to reference rex-route-helpers.js.
