+++ packages/web/src/server/routes-rex-analysis.ts
@@ -14,7 +14,7 @@ import { exec as foundationExec } from "@n-dx/llm-client";
 import type { ServerContext } from "./types.js";
 import { jsonResponse, errorResponse, readBody } from "./response-utils.js";
 import type { WebSocketBroadcaster } from "./websocket.js";
-import { insertChild, loadPRD, savePRD, appendLog } from "./routes-rex/shared.js";
+import { insertChild, loadPRD, savePRD, appendLog } from "./routes-rex/rex-route-helpers.js";
 
 import {
   type PRDItem,
