+++ packages/web/src/server/routes-rex/execution.ts
@@ -12,7 +12,7 @@ import { spawnManaged, killWithFallback, type ManagedChild } from "@n-dx/llm-cli
 import type { ServerContext } from "../types.js";
 import { jsonResponse, errorResponse, readBody } from "../response-utils.js";
 import type { WebSocketBroadcaster } from "../websocket.js";
-import { findItemById, loadPRD, appendLog } from "./shared.js";
+import { findItemById, loadPRD, appendLog } from "./rex-route-helpers.js";
 
 import {
   computeStats,
