+++ packages/web/src/server/routes-rex/prune.ts
@@ -11,7 +11,7 @@ import { join } from "node:path";
 import type { ServerContext } from "../types.js";
 import { jsonResponse, errorResponse, readBody } from "../response-utils.js";
 import type { WebSocketBroadcaster } from "../websocket.js";
-import { findItemById, loadPRD, savePRD, appendLog, parentIdOf } from "./shared.js";
+import { findItemById, loadPRD, savePRD, appendLog, parentIdOf } from "./rex-route-helpers.js";
 
 import {
   type PRDItem,
